refactor: move authentication-related models to a single folder
