+++ src/data/protocols/db/GetAccountByEmailRepository.d.ts
@@ -1,4 +1,4 @@
-import AccountModel from '../../../domain/models/Account'
+import AccountModel from '../../../domain/models/authentication/Account'
 
 export default interface GetAccountByEmailRepository {
   getByEmail: (email: string) => Promise<AccountModel | null>
