+++ src/infra/db/mongodb/account-repository/AccountMongoRepository.ts
@@ -1,6 +1,6 @@
 import AddAccountRepository from '../../../../data/protocols/db/AddAccountRepository'
 import { AddAccountModel } from '../../../../domain/usecases/authentication/AddAccount'
-import AccountModel from '../../../../domain/models/Account'
+import AccountModel from '../../../../domain/models/authentication/Account'
 import MongoHelper from '../helpers/MongoHelper'
 import GetAccountByEmailRepository from '../../../../data/protocols/db/GetAccountByEmailRepository'
